+++ lib/rubocop/cop/method_complexity_patch.rb
